Factor out various helpers for variables
